Feature/error handling

See merge request sw_track/class_06/final_project/team05/backend!16

+++ core/src/main/kotlin/elice/team5th/domain/review/dto/CreateReviewDTO.kt
@@ -3,5 +3,5 @@ package elice.team5th.domain.review.dto
 data class CreateReviewDTO(
     val contentId: Long,
     val detail: String,
-    val rating: Double
+    val rating: Int
 )
